Different changes in the client functionalities. Added log table with services not running in the infrastructure.

+++ ClientUpdater/src/clientupdater/Main.java
@@ -6,6 +6,7 @@
 package clientupdater;
 
 import java.io.BufferedReader;
+import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
@@ -13,14 +14,18 @@ import java.io.FileReader;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
+import java.io.PrintStream;
 import java.io.PrintWriter;
-import java.net.ServerSocket;
+import java.math.BigInteger;
 import java.net.Socket;
+import java.security.MessageDigest;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Date;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipInputStream;
 
 /**
  * Main class of UnaCloud Client Updater. This component is responsible for downloading the latests version of unacloud client from unacloud server instance.
@@ -34,17 +39,20 @@ public class Main {
     final static File versions = new File("versions.txt");
     static int CLIENT_PORT = 81;
     static String SERVER_IP = "";
-    public static void main(String... args){
-        print("InicioActualizador "+Arrays.toString(args)+" "+new Date(System.currentTimeMillis())+" "+new File("./").getAbsolutePath());
+    public static void main(String... args) throws FileNotFoundException{
+        if(args.length>0&&args[0]!=null&&args[0].length()<2)System.setOut(new PrintStream("logActualizador"+args[0]));
+        else System.setOut(new PrintStream("logActualizador"));
+        System.out.println("InicioActualizador "+Arrays.toString(args)+" "+new Date(System.currentTimeMillis())+" "+new File("./").getAbsolutePath());
         VariableManager.init("./vars");
         SERVER_PORT=VariableManager.getIntValue("VERSION_MANAGER_PORT");
         CLIENT_PORT = VariableManager.getIntValue("CLOUDER_CLIENT_PORT");
         SERVER_IP = VariableManager.getStringValue("CLOUDER_SERVER_IP");
-        print("InicioActualizador "+Arrays.toString(args));
+        System.out.println("InicioActualizador "+Arrays.toString(args));
         if(args.length>=1){
             int e = Integer.parseInt(args[0]);
             startClient(e);
         }
+        System.out.close();
     }
 
     /**
@@ -52,7 +60,7 @@ public class Main {
      * @param opcion
      */
     public static void startClient(int opcion) {
-        print(new File("./").getAbsolutePath());
+        System.out.println(new File("./").getAbsolutePath());
         if(opcion==6){
             opcion=1;
             try {
@@ -62,25 +70,11 @@ public class Main {
             }
         }
         if(opcion==1)doUpdate();
-        print("Actualizado");
+        System.out.println("Actualizado");
         try {
             Process p = Runtime.getRuntime().exec("java -jar ClouderClient.jar "+opcion+"\"");
         } catch (Throwable t) {
-            print("EXE: "+t.getMessage());
-        }
-    }
-
-    /**
-     * Print a string to a log file
-     * @param h
-     */
-    public static void print(String h) {
-        System.out.println(h);
-        try {
-            PrintWriter pwerr = new PrintWriter(new FileOutputStream("./Errores.txt", true), true);
-            pwerr.println(h);
-            pwerr.close();
-        } catch (FileNotFoundException ex) {
+            System.out.println("EXE: "+t.getMessage());
         }
     }
 
@@ -90,28 +84,18 @@ public class Main {
      * @param tam
      * @param isr
      */
-    private static void recibirArchivo(File ar, long tam, InputStream isr) {
-        int t;
+    private static void recibirArchivo(File ar, InputStream isr) {
         byte[] buffer = new byte[1024];
         try {
-            long cent = 0;
+            ByteArrayOutputStream baos=new ByteArrayOutputStream();
+            for (int n;(n=isr.read(buffer))!=-1;)baos.write(buffer,0,n);
+            baos.close();
             ar.getParentFile().mkdirs();
             FileOutputStream fos = new FileOutputStream(ar);
-            System.out.println(ar);
-            for (t = 0; cent + buffer.length < tam;) {
-                t = isr.read(buffer);
-                fos.write(buffer,0,t);
-                cent += t;
-            }
-            while(tam - cent != 0) {
-                int a = isr.read(buffer, 0, (int) (tam - cent));
-                fos.write(buffer, 0, (int) (tam - cent));
-                cent+=a;
-            }
-            print(ar+" "+tam+" "+cent);
+            fos.write(baos.toByteArray());
             fos.close();
         } catch (IOException ex) {
-            print(ex.getMessage());
+            System.out.println(ex.getMessage());
         }
     }
 
@@ -119,10 +103,7 @@ public class Main {
      * This method checks the current version and compares it with the version stored on unacloud server. If the version doesn't match then the newest version is downloaded and installed.
      */
     private static void doUpdate() {
-        ServerSocket ss = null;
         try {
-            //imprimir("paso 1");
-            ss = new ServerSocket(CLIENT_PORT);
             //imprimir("paso 2");
             Socket s = new Socket(SERVER_IP, SERVER_PORT);
             BufferedReader br = new BufferedReader(new InputStreamReader(s.getInputStream()));
@@ -136,57 +117,36 @@ public class Main {
             if (h.equals("version ok")) {
             } else if (h.equals("no last version")) {
                 for (int e = 1; e < versionsFile.size(); e++) {
-                    File c = new File(root, versionsFile.get(e));
-                    if (c.exists()) {
-                        c.delete();
-                    }
+                    File c = new File(versionsFile.get(e));
+                    if (c.exists())c.delete();
                 }
                 try {
-                    Socket canalArchivos = ss.accept();
-                    PrintWriter versionFile = new PrintWriter(versions);
+                    Socket canalArchivos = new Socket(SERVER_IP,SERVER_PORT);
+                    PrintWriter pwCanalArchivos=new PrintWriter(canalArchivos.getOutputStream(),true);
+                    pwCanalArchivos.println("files");
+                    PrintWriter versionFile = new PrintWriter(new FileOutputStream(versions),false);
                     versionFile.println(br.readLine());
-                    //imprimir("paso 5");
-                    int n = Integer.parseInt(br.readLine());
-                    //imprimir("paso 6 sarchivos");
-                    ss.close();
-                    InputStream isr = canalArchivos.getInputStream();
-                    String[] f;
-                    for (int e = 0; e < n; e++) {
-                        //imprimir("paso 7:" + e);
-                        f = br.readLine().split("\t");
-                        File ar = new File(root, f[0]);
-                        versionFile.println(f[0]);
-                        long tam = Long.parseLong(f[1]);
-                        recibirArchivo(ar, tam, isr);
+                    ZipInputStream zis=new ZipInputStream(canalArchivos.getInputStream());
+                    for(ZipEntry ze;(ze=zis.getNextEntry())!=null;){
+                        versionFile.println(ze.getName());
+                        recibirArchivo(new File(ze.getName()),zis);
                     }
-                    //imprimir("paso 8");
                     versionFile.flush();
                     versionFile.close();
+                    pwCanalArchivos.close();
                     canalArchivos.close();
-                    //imprimir("paso 9");
                 } catch (Throwable exx) {
+                    System.out.println("error Conetando a "+SERVER_IP+" en 80");
                     PrintWriter versionFile = new PrintWriter(versions);
                     versionFile.println("1.0");
                     versionFile.close();
-                    print(exx.getMessage());
+                    System.out.println(exx.getMessage());
                 }
-                System.out.println(br.readLine());
-                s.close();
             }
-
         } catch (Exception ex) {
-            print(ex.getLocalizedMessage());
-            if (ss != null) {
-                try {
-                    ss.close();
-                } catch (IOException ex1) {
-                    print(ex1.getMessage());
-                }
-            }
+            System.out.println(ex.getLocalizedMessage());
         }
-
     }
-
     /**
      * Rrturns the information contained on stored version file.
      * @return
@@ -205,9 +165,8 @@ public class Main {
         } catch (IOException ex) {
             ret = new ArrayList<String>();
             ret.add("NOVERSION");
-            print(ex.getMessage());
+            System.out.println(ex.getMessage());
         }
         return ret;
     }
-    
 }
